♻️ Refactor: 공통레이아웃 폴더화
